+++ lib/binance_client/client.rb
@@ -11,6 +11,7 @@ module BinanceClient
     api_action :order_book_depth
     api_action :sub_account_assets, args: [:email]
     api_action :sub_account_deposit_address, args: [:email, :coin]
+    api_action :sub_account_deposit_history
 
     attribute :host
     attribute :api_key