+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `exchange_info` endpoint
+- `sub_account_deposit_history` endpoint
 
 ## [1.5.0]
 ### Changed